update my element controller

+++ src/main/java/com/example/Create_Character/Controllers/ElementController.java
@@ -1,11 +1,10 @@
 package com.example.Create_Character.Controllers;
 
+import com.example.Create_Character.models.Element;
 import com.example.Create_Character.services.ElementService;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import javax.lang.model.element.Element;
 import java.util.List;
 
 @RestController
@@ -24,7 +23,7 @@ public class ElementController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Element> getElementById(@PathVariable Long id) {
+    public ResponseEntity<Element> getElementById(@PathVariable int id) {
         return ResponseEntity.ok(elementService.getElementById(id));
     }
 }
